Add casts for crypto++ v6 compatibility
In crypto++ commit 953252e44d25 ("Move from 'static' to 'enum' for class
constants") the macro CRYPTOPP_CONSTANT changed from adding a static
const int to the class to an enum, which does not take up space.

Therefore, definitions such as

    CRYPTOPP_CONSTANT(DEFAULT_KEYLENGTH = KeyBase::DEFAULT_KEYLENGTH)

now create DEFAULT_KEYLENGTH as an unnamed enum and not an int. The
result is that DEFAULT_KEYLENGTH is now no longer directly comparable
with ints, thanks to C++'s enum type conversion rules.

The fix is simple: cast to int, as was done in the crypto++ project in
the commit following the aforementioned: commit f5aa6f1f06b9 ("Cast
enums to int for comparison")

Closes: #8